backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.31.0
oauthlib==3.3.1
pyasn1-modules==0.4.2
pyparsing==3.2.5
requests==2.32.5
requests-oauthlib==2.0.0
boto3==1.40.37
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
python-jose[cryptography]==3.5.0
celery==5.5.3
redis==6.4.0
pyotp==2.9.0
qrcode[pil]==8.0
cryptography==45.0.6
